refactor: extract validation parts to separate functions
